#39 refactor : 업무 카드 모달 컴포넌트 분리
