+++ components/input/InputCard/ModalInput.tsx/index.ts
@@ -1,0 +1,5 @@
+export { default as InputCategory } from './InputCategory';
+export { default as InputTask } from './InputTask';
+export { default as InputSubTask } from './InputSubTask';
+export { default as InputKeyWord } from './InputKeyWord';
+export { default as InputEtc } from './InputEtc';
